Create Header and HeaderWithCloseButton
