+++ src/components/Header.js
@@ -1,58 +1,21 @@
 import React from 'react';
+import {View, Text} from 'react-native';
 import PropTypes from 'prop-types';
-import {
-    View, Image, TouchableOpacity, Text,
-} from 'react-native';
 import styles from '../styles/styles';
-import exitIcon from '../../assets/images/icon-x--20x20.png';
 
 const propTypes = {
-    // Title of the Header
-    title: PropTypes.string,
-
-    // Method to trigger when pressing close button of the header
-    onCloseButtonPress: PropTypes.func,
-};
-
-const defaultProps = {
-    title: '',
-    onCloseButtonPress: () => {},
+    /** Title of the Header */
+    title: PropTypes.string.isRequired,
 };
 
 const Header = props => (
-    <View style={styles.headerBar}>
-        <View style={[
-            styles.dFlex,
-            styles.flexRow,
-            styles.alignItemsCenter,
-            styles.flexGrow1,
-            styles.justifyContentBetween,
-            styles.overflowHidden,
-        ]}
-        >
-            <View style={[styles.flex1]}>
-                <Text numberOfLines={1} style={[styles.navText]}>
-                    {props.title}
-                </Text>
-            </View>
-            <View style={[styles.reportOptions, styles.flexRow]}>
-                <TouchableOpacity
-                    onPress={props.onCloseButtonPress}
-                    style={[styles.touchableButtonImage]}
-                >
-                    <Image
-                        resizeMode="contain"
-                        style={[styles.attachmentCloseIcon]}
-                        source={exitIcon}
-                    />
-                </TouchableOpacity>
-            </View>
-        </View>
+    <View style={[styles.flex1]}>
+        <Text numberOfLines={1} style={[styles.navText]}>
+            {props.title}
+        </Text>
     </View>
 );
 
-Header.propTypes = propTypes;
-Header.defaultProps = defaultProps;
 Header.displayName = 'Header';
-
+Header.propTypes = propTypes;
 export default Header;
